+++ flexiapi/config/app.php
@@ -51,6 +51,14 @@ return [
     'account_creation_token_retry_minutes' => env('APP_API_ACCOUNT_CREATION_TOKEN_RETRY_MINUTES', 60),
 
     /**
+     * CoTURN authentication in the provisioning
+     */
+    'coturn_server_host' => env('COTURN_SERVER_HOST', null),
+    'coturn_session_ttl_minutes' => (int)env('COTURN_SESSION_TTL_MINUTES', 60 * 24),
+    'coturn_static_auth_secret' => env('COTURN_STATIC_AUTH_SECRET', null),
+    'coturn_realm' => env('COTURN_REALM', null),
+
+    /**
      * External interfaces
      */
     'flexisip_pusher_path' => env('APP_FLEXISIP_PUSHER_PATH', null),